requirements.txt
requests==2.32.3
pdfplumber==0.11.4
PyPDF2==3.0.1
reportlab==4.2.5
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.9.2
openpyxl==3.1.5

